requirements.txt
streamlit==1.39.0
google-generativeai==0.8.3
pillow==10.4.0
python-dotenv==1.0.1
numpy==1.26.4
easyocr==1.7.1